backend/pyproject.toml
[project]
name = "storyengine"
version = "0.0.1"
description = "Story generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-code-sdk",
    "click>=8.2.1",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "instructor>=1.10.0",
    "loguru>=0.7.3",
    "openai>=1.99.6",
    "pydantic>=2.11.7",
    "python-socketio>=5.13.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.27.0",
    "pulumi>=3.0.0,<4.0.0",
    "pulumi-gcp>=8.0.0,<9.0.0",
    "google-cloud-storage>=3.4.0",
    "pygithub>=2.8.1",
    "gitpython>=3.1.45",
]

[dependency-groups]
dev = ["mypy>=1.17.1", "ruff>=0.12.8", "types-pyyaml>=6.0.12.20250822"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "data",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.9
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.uv.sources]
claude-code-sdk = { git = "https://github.com/anthropics/claude-code-sdk-python.git" }
